Cleaned up tests. All run except NOT

+++ test/d3ServerSpec.ts
@@ -3,20 +3,21 @@
  */
 import {expect} from 'chai';
 import Log from "../src/Util";
-import {Response} from "restify";
 import Server from "../src/rest/Server";
 
 var chai = require('chai');
 var chaihttp = require('chai-http');
 chai.use(chaihttp);
 
-describe.only("d3ServerSpec", function () {
+describe("d3ServerSpec", function () {
 
     let fs = require("fs");
     let server: Server;
 
     before("Start server", function (done) {
-        fs.unlinkSync("./cache.json");
+        if (fs.existsSync("./cache.json")) {
+            fs.unlinkSync("./cache.json");
+        }
         server = new Server(4321);
         server.start()
             .then(function (status) {
@@ -53,12 +54,10 @@ describe.only("d3ServerSpec", function () {
         return chai.request("http://localhost:4321")
             .put('/dataset/rooms')
             .attach("body", fs.readFileSync("./rooms.zip"), "rooms.zip")
-            .then(function (res: Response) {
-                Log.trace('then:');
+            .then(function (res: any) {
                 expect(res.status).to.be.equal(204);
             })
             .catch(function () {
-                Log.trace('catch:');
                 expect.fail();
             });
     });
@@ -67,12 +66,10 @@ describe.only("d3ServerSpec", function () {
         return chai.request("http://localhost:4321")
             .put('/dataset/rooms')
             .attach("body", fs.readFileSync("./rooms.zip"), "rooms.zip")
-            .then(function (res: Response) {
-                Log.trace('then:');
+            .then(function (res: any) {
                 expect(res.status).to.be.equal(201);
             })
             .catch(function () {
-                Log.trace('catch:');
                 expect.fail();
             });
     });
@@ -80,39 +77,33 @@ describe.only("d3ServerSpec", function () {
     it("Remove rooms.zip", function () {
         return chai.request("http://localhost:4321")
             .del('/dataset/rooms')
-            .then(function (res: Response) {
-                Log.trace('then:');
+            .then(function (res: any) {
                 expect(res.status).to.be.equal(204);
             })
             .catch(function () {
-                Log.trace('catch:');
                 expect.fail();
             });
     });
 
-    it("Try to remove rooms.zip again", function () {
+    it("Remove rooms.zip again", function () {
         return chai.request("http://localhost:4321")
             .del('/dataset/rooms')
             .then(function () {
-                Log.trace('then:');
                 expect.fail();
             })
             .catch(function (err:any) {
-                Log.trace('catch:');
                 expect(err.status).to.be.equal(404);
             });
     });
 
-    it.only("PUT invalid", function () {
+    it("PUT invalid", function () {
         return chai.request("http://localhost:4321")
             .put('/dataset/invalid')
             .attach("body", fs.readFileSync("./courses.zip"), "courses.zip")
             .then(function () {
-                Log.trace('then:');
                 expect.fail();
             })
             .catch(function (err: any) {
-                Log.trace('catch:');
                 expect(err.status).to.be.equal(400);
             });
     });
@@ -121,17 +112,15 @@ describe.only("d3ServerSpec", function () {
         return chai.request("http://localhost:4321")
             .put('/dataset/courses')
             .attach("body", fs.readFileSync("./courses.zip"), "courses.zip")
-            .then(function (res: Response) {
-                Log.trace('then:');
+            .then(function (res: any) {
                 expect(res.status).to.be.equal(204);
             })
             .catch(function () {
-                Log.trace('catch:');
                 expect.fail();
             });
     });
 
-    it("POST courses test", function () {
+    it("POST courses 200 test", function () {
         let queryJSONObject = {
             WHERE:{
                 GT:{
@@ -150,12 +139,52 @@ describe.only("d3ServerSpec", function () {
         return chai.request("http://localhost:4321")
             .post('/query')
             .send(queryJSONObject)
-            .then(function (res: Response) {
-                Log.trace('then:');
+            .then(function (res: any) {
+                expect(res.body).to.deep.equal({"render":"TABLE","result":[{"courses_dept":"math","courses_avg":99.78},{"courses_dept":"math","courses_avg":99.78}]});
                 expect(res.status).to.be.equal(200);
             })
             .catch(function () {
                 Log.trace('catch:');
+                expect.fail();
+            });
+    });
+
+    it("POST courses 400 test", function () {
+        let queryJSONObject: any = {
+            WHERE: {
+                OR: [
+                ]
+            },
+            OPTIONS: {
+                COLUMNS: [
+                    "courses_dept",
+                    "courses_avg"
+                ],
+
+                ORDER: "courses_dept",
+                FORM:"TABLE"
+            }
+        };
+        return chai.request("http://localhost:4321")
+            .post('/query')
+            .send(queryJSONObject)
+            .then(function () {
+                Log.trace('then:');
+                expect.fail();
+            })
+            .catch(function (err: any) {
+                Log.trace('catch:');
+                expect(err.status).to.be.equal(400);
+            });
+    });
+
+    it("Remove courses.zip", function () {
+        return chai.request("http://localhost:4321")
+            .del('/dataset/courses')
+            .then(function (res: any) {
+                expect(res.status).to.be.equal(204);
+            })
+            .catch(function () {
                 expect.fail();
             });
     });
